hamming working!!!
